Updates to doxygen comments
