+++ docs/website/source/api/reference/c++/solvers.rst
@@ -1,0 +1,8 @@
+Solvers
+========
+
+.. namespace:: black
+
+.. doxygenclass:: black::solvers::solver
+.. doxygenclass:: black::solvers::model
+.. doxygenenum:: black::solvers::option
