+++ src/lib/include/black/logic
@@ -30,7 +30,6 @@
 #include <black/internal/logic/terms.hpp>
 #include <black/internal/logic/operators.hpp>
 #include <black/internal/logic/module.hpp>
-#include <black/internal/logic/models.hpp>
 #include <black/internal/logic/semantics.hpp>
 
 #endif // BLACK_LOGIC_HPP
